+++ lib/pleroma/user/query.ex
@@ -59,7 +59,9 @@ defmodule Pleroma.User.Query do
             order_by: term(),
             select: term(),
             limit: pos_integer(),
-            actor_types: [String.t()]
+            actor_types: [String.t()],
+            birthday_day: pos_integer(),
+            birthday_month: pos_integer()
           }
           | map()
 
@@ -228,6 +230,20 @@ defmodule Pleroma.User.Query do
     query
     |> where([u], not is_nil(u.nickname))
     |> where([u], not like(u.nickname, "internal.%"))
+  end
+
+  defp compose_query({:birthday_day, day}, query) do
+    query
+    |> where([u], u.show_birthday == true)
+    |> where([u], not is_nil(u.birthday))
+    |> where([u], fragment("date_part('day', ?)", u.birthday) == ^day)
+  end
+
+  defp compose_query({:birthday_month, month}, query) do
+    query
+    |> where([u], u.show_birthday == true)
+    |> where([u], not is_nil(u.birthday))
+    |> where([u], fragment("date_part('month', ?)", u.birthday) == ^month)
   end
 
   defp compose_query(_unsupported_param, query), do: query