+++ lib/pleroma/web/api_spec/operations/pleroma_account_operation.ex
@@ -4,6 +4,7 @@
 
 defmodule Pleroma.Web.ApiSpec.PleromaAccountOperation do
   alias OpenApiSpex.Operation
+  alias OpenApiSpex.Schema
   alias Pleroma.Web.ApiSpec.AccountOperation
   alias Pleroma.Web.ApiSpec.Schemas.AccountRelationship
   alias Pleroma.Web.ApiSpec.Schemas.ApiError
@@ -108,6 +109,34 @@ defmodule Pleroma.Web.ApiSpec.PleromaAccountOperation do
       responses: %{
         200 => Operation.response("Relationship", "application/json", AccountRelationship),
         404 => Operation.response("Not Found", "application/json", ApiError)
+      }
+    }
+  end
+
+  def birthdays_operation do
+    %Operation{
+      tags: ["Retrieve account information"],
+      summary: "Birthday reminders",
+      description: "Birthday reminders about users you follow.",
+      operationId: "PleromaAPI.AccountController.birthdays",
+      parameters: [
+        Operation.parameter(
+          :day,
+          :query,
+          %Schema{type: :integer},
+          "Day of users' birthdays"
+        ),
+        Operation.parameter(
+          :month,
+          :query,
+          %Schema{type: :integer},
+          "Month of users' birthdays"
+        )
+      ],
+      security: [%{"oAuth" => ["read:accounts"]}],
+      responses: %{
+        200 =>
+          Operation.response("Accounts", "application/json", AccountOperation.array_of_accounts())
       }
     }
   end